RTDE Demo Remote Update
Now when we start the RTDE Console application and link it to the IP address of the local six axis robot we can control where to position the robot as well as read current positions in radians.

Later updates include adding function to the Gantt chart form application for dynamic joint position changes, pause, start, stop, ...etc.