Dashboard Gui Demo
Added a new project to test the translation of the RTDE Demo into a windows form. Which will give the user more options and control of the Universal Robot.  So far if you have the physical robot or URsim software running you can connect via TCP/IP and read the joint angles in real-time.

The next step of the demo includes hooking into the Dashboard port of the UR for exporting control scripts to the robot.

Also since the system is writing to an html file a Hololens could be utilized for displaying information to the user  that is wearing the headset. (More testing is required)